Fix image build in github actions
The updated Dockerfiles use go's ability to cross-compile in order to
speed up the multi-arch builds - no emulation is needed on the build
system, and the local image cache can be used for all steps before the
build. However, they require some additional build args, which the
scripts and Makefiles now provide.

However, these container builds require a newer version of podman than
is currently available on the GitHub Actions runners, to honor the
`--platform` flag inside the `FROM` instructions. So, the workflow now
uses docker there, which required many changes to the multi-arch build
scripts. The method is, regrettably, much more in the docker style now,
but it will still work with podman.

For reference, the builds require buildah v1.37.0, which is in podman
v5.1.0; github actions currently has podman v4.9.3.

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>